Move lazyFoldRight to collections-contrib
